+++ packages/react/src/canvas/plugins/index.ts
@@ -1,3 +1,4 @@
 export * from './CanvasHoveredMolecule';
 export * from './CanvasSelectionMolecule';
 export * from './CanvasPickAndPlopMolecule';
+export * from './CanvasDragAndDropMolecule';