+++ packages/react/src/index.tsx
@@ -3,6 +3,7 @@ export * from './canvas';
 export * from './component-metamodel';
 export * from './core';
 export * from './css-editing';
+export * from './drag-and-drop';
 export * from './hotkeys';
 export * from './node';
 export * from './rich-text';